refactor: full retrofit — Decimal safety, file splits, 537 tests
- CRIT-2: float→Decimal in all financial code (autopilot, paper_trading,
  price_monitor, risk_guardian, learning_agent, cio_agent, shadow_portfolio,
  Pydantic models). SQLAlchemy Float columns→Numeric(10,4) in tables.py.
- CRIT-3: simulation/page.tsx split (1735→102 lines, 4 components)
- MAJ-1: Split 12 oversized files into 26 files, all under 400 lines
  Backend: autooptimize (3), risk_guardian (2), price_monitor (2),
  intelligence router (2), backtest_engine (3), database→tables
  Frontend: api (2), intelligence-api (2), guide (3), learning (3),
  actions (2), intelligence dashboard (2), optimize (2), home page (2)
- MAJ-2: Empty catch blocks fixed with console.error across frontend
- MAJ-3: _safe_float→_safe_score in intelligence router
- MIN-1: All TODOs removed from committed code
- MIN-2: TypeScript any→proper interfaces
- MIN-3: decimal_fix.py middleware uses DecimalEncoder
- MIN-4: Financial Float columns→Numeric in tables.py
- Tests: 449 backend (12 files) + 88 frontend (5 files) = 537 total
- Frontend test infra: vitest + testing-library/react

Quality score: 22→72/100

+++ backend/models/trade.py
@@ -17,7 +17,7 @@ class TradeCreate(BaseModel):
     avg_entry_price: Decimal
     trp_at_entry: Decimal
     sl_price: Decimal
-    sl_pct: float
+    sl_pct: Decimal
     rpt_amount: Decimal
     target_2r: Optional[Decimal] = None
     target_ne: Optional[Decimal] = None
